+++ controllers/expose/ingressController/mock_AlternativeFQDNRedirector_test.go
@@ -7,7 +7,6 @@ import (
 
 	types "github.com/cloudogu/k8s-service-discovery/v2/internal/types"
 	mock "github.com/stretchr/testify/mock"
-
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )